Custom Protocol MSSP
This Protocol is designed to be used when you do not have any option to
use something else to communicate between two AVR microcontrollers. This
protocol is being developed on the Atmega328p but I recon that it should
work with other AVR microcontrollers as well.

It's main features are:

- Is simple to use.
  - Works on any pin (this version is currently
limited to C pin registers, later version might work on all pin
registers)
  - It doesn't change timers(Make sure both microcontrollers
overflow at the same rate!)
  - Doesn't require adressing.
  - Sends and
receives at the same time.
  - On startup it can wait infinitely long
for connection without pausing your program.

See
https://github.com/EldinZenderink/-AVR-MSSP---Master-Slave-Software-Protocol-/blob/master/README.md
for more information!